requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-JWT-Extended==4.2.3
Flask-CORS==3.0.10
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0
marshmallow==3.13.0
gunicorn==20.1.0
whitenoise==5.3.0
